refactor(gear-table): EditableFields.tsx を fields/ 配下に分割
551 行の junk drawer を 9 個の field コンポーネント + 共有 styles/types
+ barrel に分割。呼び出し側 (TableRow.tsx) の import は barrel 経由で
維持し、公開 API は無変更。

load-bearing な仕掛けは全て維持:
- EditableWeightField の key={unit} remount（g↔oz 切替の同期）
- EditableSeasonField の stopPropagation/preventDefault
- PRIORITY_STYLE は PrioritySelector 内に昇格（GearInfoSummary 側の
  Mondrian 縮退版とは用途が異なるため独立維持）

旧 EditableFields.tsx は後方互換用 re-export に退化 (20 行)。
新規コードは fields/ 直接参照を推奨。

検証:
- npm run typecheck 新規エラーなし
- npm run build 成功
- npm run lint 通過
