+++ client/components/GearTable/fields/index.ts
@@ -1,0 +1,16 @@
+/**
+ * GearTable の各セル用インライン編集コンポーネント群。
+ *
+ * 旧 `EditableFields.tsx`（551 行の junk drawer）を機能ごとのファイルに分割したもの。
+ * `EditableFields.tsx` は後方互換のためこの barrel を再 export する薄いファサードに退化している。
+ */
+export { EditableImageField } from './EditableImageField'
+export { EditableTextField } from './EditableTextField'
+export { EditableCategoryField } from './EditableCategoryField'
+export { EditablePriceField } from './EditablePriceField'
+export { EditableWeightField } from './EditableWeightField'
+export { EditableSeasonField } from './EditableSeasonField'
+export { QuantitySelector } from './QuantitySelector'
+export { PrioritySelector } from './PrioritySelector'
+export { EditableWeightClassField } from './EditableWeightClassField'
+export type { Currency } from './types'
